Refactor all int*4 params to Geometry
Also converge on int/uint for our public APIs, type casting in x11 package

+++ internal/x11/border.go
@@ -8,24 +8,24 @@ import (
 )
 
 // BorderWidth is the number of pixels required for a border
-func BorderWidth(win XWin) uint16 {
+func BorderWidth(win XWin) uint {
 	if !win.Properties().Decorated() {
 		return 0
 	}
-	return ScaleToPixels(wmTheme.BorderWidth, fynedesk.Instance().Screens().ScreenForWindow(win))
+	return uint(ScaleToPixels(wmTheme.BorderWidth, fynedesk.Instance().Screens().ScreenForWindow(win)))
 }
 
 // ButtonWidth is the number of pixels required for a border button
-func ButtonWidth(win XWin) uint16 {
-	return ScaleToPixels(wmTheme.ButtonWidth, fynedesk.Instance().Screens().ScreenForWindow(win))
+func ButtonWidth(win XWin) uint {
+	return uint(ScaleToPixels(wmTheme.ButtonWidth, fynedesk.Instance().Screens().ScreenForWindow(win)))
 }
 
 // TitleHeight is the number of pixels required for a title bar
-func TitleHeight(win XWin) uint16 {
-	return ScaleToPixels(wmTheme.TitleHeight, fynedesk.Instance().Screens().ScreenForWindow(win))
+func TitleHeight(win XWin) uint {
+	return uint(ScaleToPixels(wmTheme.TitleHeight, fynedesk.Instance().Screens().ScreenForWindow(win)))
 }
 
 // ScaleToPixels calculates the pixels required to show a specified Fyne dimension on the given screen
-func ScaleToPixels(i int, screen *fynedesk.Screen) uint16 {
-	return uint16(float32(i) * screen.CanvasScale())
+func ScaleToPixels(i int, screen *fynedesk.Screen) int {
+	return int(float32(i) * screen.CanvasScale())
 }
